expense_automation/requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pandas
openpyxl
python-dotenv
